+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
+## [1.3.0] - 2025-02-20
+### Added
+- AddressablesLabelUtility now has a preview of affected assets
+
 ## [1.2.0] - 2025-02-19
+
 ### Added
 - AddressablesLabelUtility can add addressable labels to assets
   - filter for assets by type and some other parameters using unity's search